pyproject.toml
[project]
name = "house-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.12.12",
    "beautifulsoup4==4.13.4",
    "dotenv>=0.9.9",
    "python-telegram-bot>=22.7",
    "requests==2.32.4",
    "typing>=3.10.0.0",
]
